+++ crates/jmap/src/registry/set.rs
@@ -722,7 +722,9 @@ impl RegistrySet for Server {
             ObjectType::AccountSettings
             | ObjectType::ApiKey
             | ObjectType::AccountPassword
-            | ObjectType::AppPassword => account_set(set).await.map(|set| set.into_response()),
+            | ObjectType::AppPassword => Box::pin(account_set(set))
+                .await
+                .map(|set| set.into_response()),
 
             ObjectType::QueuedMessage => {
                 queued_message_set(set).await.map(|set| set.into_response())